add set node by guid from url
