+++ src/app/core-components/backup-modal/export-backup-modal/export-backup-tree/export-backup-tree.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { ExportTreeNode } from '../exportTree';
+import { ExportTreeNode, updateChildren, updateParent } from '../exportTree';
 import { IconDefinition, faAngleDown, faAngleRight, faBuilding, faIndustry, faCalendar, faScrewdriverWrench } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
@@ -27,25 +27,7 @@ export class ExportBackupTreeComponent {
   }
 
   onCheckboxChange(node: ExportTreeNode) {
-    this.updateChildren(node, node.checked);
-    this.updateParent(node.parent);
-  }
-
-  updateChildren(node: ExportTreeNode, checked: boolean) {
-    // update the checked state of the node and its children recursively
-    node.checked = checked;
-    node.indeterminate = false;
-    if (node.children && node.children.length > 0) {
-      node.children.forEach(child => this.updateChildren(child, checked));
-    }
-  }
-  updateParent(parent: ExportTreeNode) {
-    if (!parent) return;
-    const children = parent.children || [];
-    const allChecked = children.every(child => child.checked);
-    const anyCheckedOrIndeterminate = children.some(child => child.checked || child.indeterminate);
-    parent.checked = allChecked;
-    parent.indeterminate = !allChecked && anyCheckedOrIndeterminate;
-    this.updateParent(parent.parent);
+    updateChildren(node, node.checked);
+    updateParent(node.parent);
   }
 }
